engines/raw-data-updater/requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
Werkzeug==0.11.3
gunicorn==19.4.5
marshmallow==2.5.0
webargs==1.2.0
py-sltc==0.8.5
sltc-orm==1.5.1
pika==0.10.0
SQLAlchemy==1.0.11
pyodbc==3.0.7
nose==1.3.3
python-dateutil==2.2
elasticsearch==5.1.0
scikit-learn==0.17.1
sltc-ml-models==1.0.4
mds-specs==1.1.4
requests==2.10.0
pymongo==3.2.2
scrapy==1.0.5
xmltodict==0.10.1
PyPDF2==1.26.0
termcolor==1.1.0
multipledispatch==0.4.8
psycopg2==2.6.2
